Fix pyinstaller bundling issue
