+++ .github/workflows/publish.yml
@@ -42,12 +42,7 @@ jobs:
         shell: bash
         run: |
           python -m pip install pyinstaller .
-          pyinstaller --name doppkit-win64-cli \
-            --onefile \
-            --console \
-            --icon src/doppkit/resources/grid-icon.ico \
-            --hidden-import aiofiles \
-            cli.py
+          pyinstaller cli.spec
       - name: Zip Output
         run: |
           &"C:/Program Files/7-Zip/7z.exe" a ./dist/doppkit-win64-cli.zip ./dist/doppkit-win64-cli.exe
