--- cli.py
@@ -1,11 +1,0 @@
-from doppkit import app
-
-
-def main():
-    """The main function for our script."""
-
-    app.cli()
-
-
-if __name__ == '__main__':
-    main()
