+++ src/doppkit/__init__.py
@@ -1,6 +1,4 @@
-__version__ = "0.1.14"
+__version__ = "0.1.15"
 
 from . import cache
 from .app import Application
-
-from .sync import sync
