--- src/doppkit/__main__.py
@@ -1,2 +1,0 @@
-from .app import cli
-from .qt.__main__ import main as gui
