+++ src/doppkit/gui/cache.py
