+++ src/doppkit/gui/window.py
